#147 move listen_to_jobs to change dir
