+++ src/real/examples/example_reals.py
@@ -279,220 +279,3 @@ def create_example_real4() -> RealUnit:
     # music_real.set_all_econunits_role(zia_text)
 
     return music_real
-
-
-def casa_text() -> str:
-    return "casa"
-
-
-def cook_text() -> str:
-    return "cook"
-
-
-def eat_text() -> str:
-    return "eat"
-
-
-def hungry_text() -> str:
-    return "hungry"
-
-
-def full_text() -> str:
-    return "full"
-
-
-def sanitation_text():
-    return "sanitation"
-
-
-def clean_text():
-    return "clean"
-
-
-def dirty_text():
-    return "dirty"
-
-
-def sweep_text():
-    return "sweep"
-
-
-def run_text():
-    return "run"
-
-
-def casa_road() -> RoadUnit:
-    return create_road(get_default_real_id_roadnode(), casa_text())
-
-
-def cook_road() -> RoadUnit:
-    return create_road(casa_road(), cook_text())
-
-
-def eat_road() -> RoadUnit:
-    return create_road(casa_road(), eat_text())
-
-
-def hungry_road() -> RoadUnit:
-    return create_road(eat_road(), hungry_text())
-
-
-def full_road() -> RoadUnit:
-    return create_road(eat_road(), full_text())
-
-
-def sanitation_road() -> RoadUnit:
-    return create_road(casa_road(), sanitation_text())
-
-
-def clean_road() -> RoadUnit:
-    return create_road(sanitation_road(), clean_text())
-
-
-def dirty_road() -> RoadUnit:
-    return create_road(sanitation_road(), dirty_text())
-
-
-def sweep_road() -> RoadUnit:
-    return create_road(casa_road(), sweep_text())
-
-
-def run_road() -> RoadUnit:
-    return create_road(casa_road(), run_text())
-
-
-def get_example_yao_agenda() -> AgendaUnit:
-    yao_text = "Yao"
-    zia_text = "Zia"
-    bob_text = "Bob"
-    yao_speaker = agendaunit_shop(yao_text, get_default_real_id_roadnode())
-    yao_speaker.add_idea(ideaunit_shop(run_text()), casa_road())
-    yao_speaker.add_partyunit(yao_text, debtor_weight=10)
-    yao_speaker.add_partyunit(zia_text, debtor_weight=30)
-    yao_speaker.add_partyunit(bob_text, debtor_weight=40)
-    yao_speaker.set_party_pool(80)
-    return yao_speaker
-
-
-def get_example_yao_job1_speaker() -> AgendaUnit:
-    yao_text = "Yao"
-    yao_speaker = get_example_yao_agenda()
-    yao_speaker.set_party_pool(40)
-    yao_speaker.add_idea(ideaunit_shop(cook_text(), pledge=True), casa_road())
-    yao_speaker.add_idea(ideaunit_shop(hungry_text()), eat_road())
-    yao_speaker.add_idea(ideaunit_shop(full_text()), eat_road())
-    cook_ideaunit = yao_speaker.get_idea_obj(cook_road())
-    cook_ideaunit._assignedunit.set_suffgroup(yao_text)
-    yao_speaker.edit_reason(cook_road(), eat_road(), hungry_road())
-    yao_speaker.set_belief(eat_road(), hungry_road())
-    return yao_speaker
-
-
-def get_example_yao_job2_speaker() -> AgendaUnit:
-    yao_text = "Yao"
-    yao_speaker = get_example_yao_agenda()
-    yao_speaker.set_party_pool(30)
-    yao_speaker.add_idea(ideaunit_shop(cook_text(), pledge=True), casa_road())
-    yao_speaker.add_idea(ideaunit_shop(hungry_text()), eat_road())
-    yao_speaker.add_idea(ideaunit_shop(full_text()), eat_road())
-    cook_ideaunit = yao_speaker.get_idea_obj(cook_road())
-    cook_ideaunit._assignedunit.set_suffgroup(yao_text)
-    yao_speaker.edit_reason(cook_road(), eat_road(), hungry_road())
-    yao_speaker.set_belief(eat_road(), hungry_road())
-
-    yao_speaker.add_idea(ideaunit_shop(sweep_text(), pledge=True), casa_road())
-    yao_speaker.add_idea(ideaunit_shop(dirty_text()), sanitation_road())
-    yao_speaker.add_idea(ideaunit_shop(clean_text()), sanitation_road())
-    yao_speaker.edit_reason(sweep_road(), sanitation_road(), dirty_road())
-    yao_speaker.set_belief(sweep_road(), dirty_road())
-    return yao_speaker
-
-
-def get_example_yao_job3_speaker() -> AgendaUnit:
-    yao_speaker = get_example_yao_agenda()
-    yao_speaker.set_party_pool(10)
-    yao_speaker.add_idea(ideaunit_shop(sweep_text(), pledge=True), casa_road())
-    yao_speaker.add_idea(ideaunit_shop(dirty_text()), sanitation_road())
-    yao_speaker.add_idea(ideaunit_shop(clean_text()), sanitation_road())
-    yao_speaker.edit_reason(sweep_road(), sanitation_road(), dirty_road())
-    yao_speaker.set_belief(sweep_road(), dirty_road())
-    return yao_speaker
-
-
-def get_usa_road() -> RoadUnit:
-    return create_road(get_default_real_id_roadnode(), "USA")
-
-
-def get_iowa_text() -> RoadNode:
-    return "Iowa"
-
-
-def get_ohio_text() -> RoadNode:
-    return "Ohio"
-
-
-def get_utah_text() -> RoadNode:
-    return "Utah"
-
-
-def get_iowa_road() -> RoadUnit:
-    return create_road(get_usa_road(), "Iowa")
-
-
-def get_ohio_road() -> RoadUnit:
-    return create_road(get_usa_road(), "Ohio")
-
-
-def get_utah_road() -> RoadUnit:
-    return create_road(get_usa_road(), "Utah")
-
-
-def get_yao_ohio_agendahub() -> AgendaHub:
-    yao_agenda = get_example_yao_agenda()
-    return agendahub_shop(
-        reals_dir=get_test_reals_dir(),
-        real_id=yao_agenda._real_id,
-        person_id=yao_agenda._owner_id,
-        econ_road=get_ohio_road(),
-        nox_type=pipeline_duty_work_text(),
-    )
-
-
-def get_yao_iowa_agendahub() -> AgendaHub:
-    yao_agenda = get_example_yao_agenda()
-    return agendahub_shop(
-        reals_dir=get_test_reals_dir(),
-        real_id=yao_agenda._real_id,
-        person_id=yao_agenda._owner_id,
-        econ_road=get_iowa_road(),
-        nox_type=pipeline_duty_work_text(),
-    )
-
-
-def get_zia_utah_agendahub() -> AgendaHub:
-    yao_agenda = get_example_yao_agenda()
-    return agendahub_shop(
-        reals_dir=get_test_reals_dir(),
-        real_id=yao_agenda._real_id,
-        person_id="Zia",
-        econ_road=get_utah_road(),
-        nox_type=pipeline_duty_work_text(),
-    )
-
-
-def get_example_yao_duty_with_3_healers():
-    yao_duty = get_example_yao_agenda()
-    yao_text = yao_duty.get_party("Yao").party_id
-    bob_text = yao_duty.get_party("Bob").party_id
-    zia_text = yao_duty.get_party("Zia").party_id
-    iowa_idea = ideaunit_shop(get_iowa_text(), _problem_bool=True)
-    ohio_idea = ideaunit_shop(get_ohio_text(), _problem_bool=True)
-    utah_idea = ideaunit_shop(get_utah_text(), _problem_bool=True)
-    iowa_idea._healerhold.set_group_id(get_yao_iowa_agendahub().person_id)
-    ohio_idea._healerhold.set_group_id(get_yao_ohio_agendahub().person_id)
-    utah_idea._healerhold.set_group_id(get_zia_utah_agendahub().person_id)
-    yao_duty.add_idea(iowa_idea, get_usa_road())
-    yao_duty.add_idea(ohio_idea, get_usa_road())
-    yao_duty.add_idea(utah_idea, get_usa_road())
-
-    return yao_duty
